+++ src/main.ts
@@ -1,12 +1,18 @@
 import { ConfigService } from '@nestjs/config'
 import { NestFactory } from '@nestjs/core'
+import type { NestFastifyApplication } from '@nestjs/platform-fastify'
 
 import { AppModule } from '~/app.module'
+import { setupFastifyAdapter } from '~/helpers/setup/fastify-adapter'
 import { setupSwagger } from '~/helpers/setup/swagger'
 import { setupValidation } from '~/helpers/setup/validation'
 
 const bootstrap = async () => {
-  const app = await NestFactory.create(AppModule)
+  const app = await NestFactory.create<NestFastifyApplication>(
+    AppModule,
+    setupFastifyAdapter(),
+  )
+
   const configService = app.get<ConfigService>(ConfigService)
 
   setupSwagger('swagger', app)